+++ package.json
@@ -14,11 +14,14 @@
 		"node": ">=8"
 	},
 	"scripts": {
-		"test": "xo && ava && tsd"
+		"test": "xo && npm run build && ava",
+		"build": "del-cli dist && tsc",
+		"prepack": "npm run build"
 	},
+	"main": "dist",
 	"files": [
-		"index.js",
-		"index.d.ts"
+		"dist/index.js",
+		"dist/index.d.ts"
 	],
 	"keywords": [
 		"memoize",
@@ -38,10 +41,12 @@
 		"mimic-fn": "^3.0.0"
 	},
 	"devDependencies": {
+		"@sindresorhus/tsconfig": "^0.7.0",
 		"ava": "^2.4.0",
+		"del-cli": "^3.0.1",
 		"delay": "^4.1.0",
 		"serialize-javascript": "^2.1.0",
-		"tsd": "^0.11.0",
+		"typescript": "^4.0.3",
 		"xo": "^0.25.3"
 	}
 }
