+++ tsconfig.json
@@ -1,0 +1,9 @@
+{
+	"extends": "@sindresorhus/tsconfig",
+	"compilerOptions": {
+		"outDir": "dist"
+	},
+	"files": [
+		"index.ts",
+	]
+}
